+++ tinycontent/migrations/0003_update_slug.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import models, migrations
+import autoslug.fields
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('tinycontent', '0002_tinycontentfileupload'),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name='tinycontentfileupload',
+            name='slug',
+            field=autoslug.fields.AutoSlugField(populate_from=b'name', unique=True, editable=False),
+        ),
+    ]